refactor: remove keyboardIcons and ChevronLeftIcon.tsx, right, top icons. Replace them in components
