+++ src/shared/assets/icons/index.ts
@@ -4,15 +4,12 @@ export { default as BookmarkIcon } from './BookmarkIcon'
 export { default as BookmarkOutlineIcon } from './BookmarkOutlineIcon'
 export { default as CalendarIcon } from './CalendarIcon'
 export { default as CalendarOutlineIcon } from './CalendarOutlineIcon'
-export { default as ChevronLeftIcon } from './ChevronLeftIcon'
-export { default as ChevronRightIcon } from './ChevronRightIcon'
 export { default as CloseIcon } from './CloseIcon'
+export { default as ChevronIcon } from './ChevronIcon'
 export { default as EyeIcon } from './EyeIcon'
 export { default as EyeOffIcon } from './EyeOffIcon'
 export { default as HomeIcon } from './HomeIcon'
 export { default as HomeOutlineIcon } from './HomeOutlineIcon'
-export { default as KeyboardArrowLeft } from './KeyboardArrowLeft'
-export { default as KeyboardArrowRight } from './KeyboardArrowRight'
 export { default as LogOutIcon } from './LogOutIcon'
 export { default as MessageIcon } from './MessageIcon'
 export { default as MessageOutlineIcon } from './MessageOutlineIcon'
