+++ src/widgets/sidebar/ui/desktop-sidebar/toggle-collapsed-buttons.tsx
@@ -1,4 +1,4 @@
-import { ChevronLeftIcon, ChevronRightIcon } from '@/shared/assets/icons'
+import { ChevronIcon } from '@/shared/assets/icons'
 import { useLayoutContext } from '@/shared/layouts/context/layout-context'
 import { ReturnComponent } from '@/shared/types'
 import { Button } from '@/shared/ui/button'
@@ -6,23 +6,27 @@ import { Button } from '@/shared/ui/button'
 export const ToggleCollapsedButtons = (): ReturnComponent => {
   const { isCollapsed, toggleSidebar } = useLayoutContext()
 
+  const classes = {
+    button: `top-1/5 fixed z-10 !h-8 !w-8 rounded-full !bg-Dark-100/25`,
+  }
+
   return (
     <>
       {isCollapsed ? (
         <Button
-          className={'top-1/5 fixed left-[30px] z-10 !h-8 !w-8 rounded-full !bg-Dark-100/25 '}
+          className={`${classes.button} left-[30px]`}
           onClick={toggleSidebar}
           variant={'text'}
         >
-          <ChevronRightIcon className={'!fill-Light-100/70'} />
+          <ChevronIcon className={'h-[16px] w-[16px] rotate-90'} />
         </Button>
       ) : (
         <Button
-          className={'top-1/5 fixed left-[200px] z-10 !h-8 !w-8 rounded-full !bg-Dark-100/25 '}
+          className={`${classes.button} left-[200px]`}
           onClick={toggleSidebar}
           variant={'text'}
         >
-          <ChevronLeftIcon className={'!fill-Light-100/70'} />
+          <ChevronIcon className={'h-[16px] w-[16px] -rotate-90'} />
         </Button>
       )}
     </>
